UI fixes + instructions on welcome page
Fixed snackbar: now is always centered;

If an error occurs during boot the error is now displayed on the loading page, useful for mobile users when having issues;

Add a message that is displayed to users with no heroes in their rosters. Onclick it shows some quick instructions on how to use the website;

Fixed toolbar's hover effect hiding the active icon green background.